Fix java coding standard violations.

+++ src/main/java/seedu/address/logic/commands/DeleteTagCommand.java
@@ -50,6 +50,11 @@ public class DeleteTagCommand extends Command {
         }
 
         Person deleteFromPerson = lastShownList.get(targetIndex.getZeroBased());
+        //check if the tag with the given name exist
+        if (!deleteFromPerson.getTags().contains(tagToDelete)) {
+            throw new CommandException(Messages.MESSAGE_UNEXISTING_TAG);
+        }
+
         model.deleteTag(deleteFromPerson, tagToDelete);
         return new CommandResult(String.format(MESSAGE_SUCCESS, deleteFromPerson));
     }
@@ -58,7 +63,8 @@ public class DeleteTagCommand extends Command {
     public boolean equals(Object other) {
         return other == this // short circuit if same object
                 || (other instanceof DeleteTagCommand // instanceof handles nulls
-                && targetIndex.equals(((DeleteTagCommand) other).targetIndex) && tagToDelete.equals(((DeleteTagCommand) other).tagToDelete));
+                && targetIndex.equals(((DeleteTagCommand) other).targetIndex)
+                        && tagToDelete.equals(((DeleteTagCommand) other).tagToDelete));
     }
 
 }
